Move writer integration tests for indentation to tests/writer-indentation.rs
